Find default sink name upon connect
The sink name in src/sinks.ini didn't work on all devices, and it's
pretty trivial to get the default sink name, so let's just discover the
sink name as opposed to hardcoding it.

It seems that the default_sink_name property in the status applet
private structure is unused, so perhaps we ought to remove it.